backend/openai_service/test_openai.py
from openai import OpenAI
from dotenv import load_dotenv
import os

load_dotenv()  

client = OpenAI(api_key=os.getenv("OPENAI_API_KEY"))

def test():
    res = client.chat.completions.create(
        model="gpt-3.5-turbo",
        messages=[{"role": "user", "content": "Hi"}]
    )
    print(res.choices[0].message.content)

test()