backend/requirements.txt
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
httplib2==0.22.0
idna==3.10
oauth2client==4.1.3
oauthlib==3.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dotenv==1.1.1
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
urllib3==2.5.0
fastapi
uvicorn
jinja2
python-multipart
itsdangerous
psycopg2-binary
sqlalchemy
python-jose[cryptography]
backoff>=2.2.1
openai
rapidfuzz
fuzzywuzzy
bs4